requirements.txt
absl-py==2.1.0
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.5.0
APScheduler==3.10.4
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.1.0
contourpy==1.2.0
cycler==0.12.1
datatile==1.0.3
Django==3.2.16
docs==0.1.0
docx==0.2.4
flake8==6.1.0
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.3.3
fuzzywuzzy==0.18.0
gast==0.5.4
git-filter-repo==2.45.0
google-pasta==0.2.0
grpcio==1.64.0
h5py==3.11.0
idna==3.4
iniconfig==2.0.0
isoweek==1.3.3
joblib==1.3.2
keras==3.3.3
kiwisolver==1.4.5
Levenshtein==0.25.1
libclang==18.1.1
logging==0.4.9.6
lxml==4.9.3
magic-filter==1.0.11
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.4
mypy==1.7.1
mypy-extensions==1.0.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.2.1
pandas-summary==0.2.0
pillow==10.2.0
pluggy==1.4.0
protobuf==4.25.3
py==1.11.0
pycodestyle==2.11.1
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.1.0
Pygments==2.18.0
pyparsing==3.1.2
pytest==7.1.2
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
pytz==2022.7.1
rapidfuzz==3.9.1
requests==2.31.0
rich==13.7.1
schedule==1.2.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sklearn-pandas==2.2.0
soupsieve==2.4.1
sqlparse==0.4.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.4.0
tomli==2.0.1
traceml==1.0.8
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==5.0.1
urllib3==2.0.3
Werkzeug==3.0.3
wrapt==1.16.0
xgboost==2.0.3
yarl==1.8.2
